Move animation timestamp into the animation module
