+++ Animation/Animation/Timestamp.cpp
@@ -1,0 +1,50 @@
+#include <AnimationStdAfx.h>
+
+#include <Animation/Timestamp.h>
+
+namespace GLEngine::Animation {
+
+//=================================================================================
+// S_Timestamp
+//=================================================================================
+S_Timestamp::S_Timestamp(float timestamp)
+	: m_Timestamp(S_Timestamp::Clamp(timestamp))
+{
+}
+
+//=================================================================================
+S_Timestamp::S_Timestamp()
+	: m_Timestamp(0.f)
+{
+}
+
+//=================================================================================
+float S_Timestamp::Clamp(float f)
+{
+	return std::clamp(f, 0.0f, 1.0f);
+}
+
+//=================================================================================
+bool S_Timestamp::operator>=(const S_Timestamp& other) const
+{
+	return m_Timestamp >= other.m_Timestamp;
+}
+
+//=================================================================================
+bool S_Timestamp::operator>(const S_Timestamp& other) const
+{
+	return m_Timestamp > other.m_Timestamp;
+}
+
+//=================================================================================
+S_Timestamp S_Timestamp::operator-(const S_Timestamp& rhs) const
+{
+	return S_Timestamp(m_Timestamp - rhs.m_Timestamp);
+}
+
+//=================================================================================
+S_Timestamp S_Timestamp::operator+(const S_Timestamp& rhs) const
+{
+	return S_Timestamp(m_Timestamp + rhs.m_Timestamp);
+}
+}
