+++ Renderer/Renderer/Animation/AnimationStructures.h
@@ -2,33 +2,20 @@
 
 #include <Renderer/RendererApi.h>
 
+#include <Animation/Timestamp.h>
+
 #include <glm/gtc/quaternion.hpp>
 
 namespace GLEngine::Renderer::Animation {
 
-struct S_Timestamp {
-	S_Timestamp();
-	RENDERER_API_EXPORT S_Timestamp(float timestamp);
-
-	[[nodiscard]] bool		  operator>=(const S_Timestamp& other) const;
-	[[nodiscard]] bool		  operator>(const S_Timestamp& other) const;
-	[[nodiscard]] S_Timestamp operator-(const S_Timestamp& rhs) const;
-	[[nodiscard]] S_Timestamp operator+(const S_Timestamp& rhs) const;
-	[[nodiscard]] float		  GetValue() const { return m_Timestamp; }
-
-private:
-	static float Clamp(float f);
-	float		 m_Timestamp;
-};
-
 struct S_BoneKeyframe {
 	S_BoneKeyframe() = default;
-	S_BoneKeyframe(const glm::mat4& matrix, S_Timestamp timestamp);
-	S_BoneKeyframe(const glm::vec3& transformation, const glm::quat& rotation, S_Timestamp timestamp);
-	RENDERER_API_EXPORT glm::mat4 GetTransformationMatrix() const;
-	S_Timestamp					  m_Timestamp;
-	glm::vec3					  m_Transform;
-	glm::quat					  m_Rotation;
+	S_BoneKeyframe(const glm::mat4& matrix, GLEngine::Animation::S_Timestamp timestamp);
+	S_BoneKeyframe(const glm::vec3& transformation, const glm::quat& rotation, GLEngine::Animation::S_Timestamp timestamp);
+	RENDERER_API_EXPORT glm::mat4	 GetTransformationMatrix() const;
+	GLEngine::Animation::S_Timestamp m_Timestamp;
+	glm::vec3						 m_Transform;
+	glm::quat						 m_Rotation;
 };
 
 // todo: move ctor
@@ -38,7 +25,7 @@ public:
 	explicit C_BoneTimeline(std::size_t numTimestamps);
 
 	void						 AddBoneKeyFrame(std::size_t index, S_BoneKeyframe&& keyframe);
-	[[nodiscard]] S_BoneKeyframe Sample(S_Timestamp timestamp) const;
+	[[nodiscard]] S_BoneKeyframe Sample(GLEngine::Animation::S_Timestamp timestamp) const;
 
 private:
 	// ideally BST because intervals do not overlap
