+++ Renderer/Renderer/Animation/SkeletalAnimation.cpp
@@ -19,7 +19,7 @@ void C_SkeletalAnimation::SetBoneTimeline(std::size_t index, C_BoneTimeline&& ti
 }
 
 //=================================================================================
-C_Pose C_SkeletalAnimation::GetPose(const S_Timestamp timestamp) const
+C_Pose C_SkeletalAnimation::GetPose(const GLEngine::Animation::S_Timestamp timestamp) const
 {
 	std::vector<S_BoneKeyframe> ret;
 	ret.resize(m_Timelines.size());
